prompts_completion_maker/pyproject.toml
[tool.poetry]
name = "fine-tuning-demo"
version = "0.1.0"
description = ""
authors = ["yusuf-74 <yachour31@gmail.com>"]
readme = "README.md"
packages = [{include = "fine_tuning_demo"}]

[tool.poetry.dependencies]
python = "^3.10"
openai = "^0.27.1"
python-decouple = "^3.8"


[tool.poetry.group.dev.dependencies]
autopep8 = "^2.0.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
